Move big query lib into its own module
